mrc-misc-fn-modules(extract-msprobe-move,mechanical_provable): move msprobe debugger init to msprobe module (cut+paste)
