+++ python/sglang/srt/model_executor/model_runner.py
@@ -134,6 +134,9 @@ from sglang.srt.model_executor.model_runner_components.load_model_utils import (
     report_online_quantization,
     resolve_sliding_window_size,
 )
+from sglang.srt.model_executor.model_runner_components.misc_utils import (
+    create_msprobe_debugger,
+)
 from sglang.srt.model_executor.model_runner_components.ngram_embedding_manager import (
     NgramEmbeddingManager,
 )
@@ -2032,29 +2035,3 @@ class ModelRunner(ModelRunnerKVCacheMixin):
             load_format=load_format,
         )
         self.load_config = load_config
-
-
-import logging
-from typing import TYPE_CHECKING, Any, Optional
-
-if TYPE_CHECKING:
-    from sglang.srt.server_args import ServerArgs
-
-logger = logging.getLogger(__name__)
-
-
-def create_msprobe_debugger(server_args: ServerArgs) -> Optional[Any]:
-    if server_args.msprobe_dump_config is None:
-        return None
-
-    try:
-        from msprobe.pytorch import PrecisionDebugger, seed_all
-    except ImportError:
-        logger.warning(
-            "Please install msprobe for tensor data dump: pip install mindstudio-probe --pre, "
-            "see https://gitcode.com/Ascend/msprobe for details."
-        )
-        return None
-
-    seed_all(mode=True)
-    return PrecisionDebugger(config_path=server_args.msprobe_dump_config)
