+++ python/sglang/srt/model_executor/model_runner_components/misc_utils.py
@@ -1,0 +1,26 @@
+from __future__ import annotations
+
+import logging
+from typing import TYPE_CHECKING, Any, Optional
+
+if TYPE_CHECKING:
+    from sglang.srt.server_args import ServerArgs
+
+logger = logging.getLogger(__name__)
+
+
+def create_msprobe_debugger(server_args: ServerArgs) -> Optional[Any]:
+    if server_args.msprobe_dump_config is None:
+        return None
+
+    try:
+        from msprobe.pytorch import PrecisionDebugger, seed_all
+    except ImportError:
+        logger.warning(
+            "Please install msprobe for tensor data dump: pip install mindstudio-probe --pre, "
+            "see https://gitcode.com/Ascend/msprobe for details."
+        )
+        return None
+
+    seed_all(mode=True)
+    return PrecisionDebugger(config_path=server_args.msprobe_dump_config)
